LayoutTests/platform/ios-simulator/editing/pasteboard/copy-paste-first-line-in-textarea-expected.txt


FAIL selectFirstLineIncludingLineBreak(); copyPaste(); textarea.selectionEnd should be 6. Was 0.
